Improve the judgment on whether rknpu needs to be started

+++ scripts/firstboot/firstboot_noble_stress-test.sh
@@ -351,15 +351,13 @@ function clean_debootstrap_dir() {
 }
 
 function enable_rknpu() {
-	if modinfo rknpu 2>/dev/null;then
+	if [ -f "/usr/local/lib/systemd/system/rknn.service" ] && modinfo rknpu 2>/dev/null;then
 		echo rknpu > /etc/modules-load.d/rknpu.conf
 		echo "alias rknpu rknpu" > /etc/modprobe.d/rknpu.conf
 		modprobe rknpu
-		if [ -f "/usr/local/lib/systemd/system/rknn.service" ];then
-			ldconfig
-			systemctl enable rknn.service
-			systemctl start rknn.service
-		fi
+		ldconfig
+		systemctl enable rknn.service
+		systemctl start rknn.service
 	fi
 }
